+++ package.json
@@ -28,6 +28,7 @@
     "react-native-svg-transformer": "^1.5.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-vector-icons": "^10.2.0",
+    "react-native-video": "^6.10.0",
     "react-redux": "^9.2.0"
   },
   "devDependencies": {
