+++ src/assets/svgs/index.ts
@@ -8,3 +8,9 @@ export {default as HomeFilled} from './home-filled.svg';
 export {default as PostOutLined} from './post-outlined.svg';
 export {default as PostFilled} from './post-filled.svg';
 export {default as AddPostIcon} from './add-post.svg';
+export {default as UploadIcon} from './upload.svg';
+export {default as GalleryIcon} from './gallery.svg';
+export {default as CameraIcon} from './camera.svg';
+export {default as DeleteIcon} from './delete.svg';
+export {default as PlayIcon} from './playicon.svg';
+export {default as PauseIcon} from './pauseicon.svg';
