--- src/screens/posts/postCreateStyle.ts
@@ -1,62 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-  modalBackground: {
-    flex: 1,
-    backgroundColor: 'rgba(0,0,0,0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  modalContainer: {
-    width: '90%',
-    backgroundColor: 'white',
-    padding: 20,
-    borderRadius: 10,
-    alignItems: 'center',
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  errorText: {
-    color: 'red',
-    alignSelf: 'stretch',
-  },
-  mediaButtonsContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 10,
-  },
-  mediaPreviewContainer: {
-    marginTop: 10,
-    marginRight: 10,
-    alignItems: 'center',
-  },
-  mediaPreview: {
-    width: 100,
-    height: 100,
-    borderRadius: 5,
-  },
-  submitButton: {
-    backgroundColor: '#007BFF',
-    padding: 10,
-    borderRadius: 5,
-    marginTop: 10,
-    alignItems: 'center',
-    width: '100%',
-  },
-  submitButtonText: {
-    color: '#fff',
-    fontWeight: 'bold',
-  },
-  closeButton: {
-    marginTop: 10,
-    alignSelf: 'center',
-  },
-  closeButtonText: {
-    color: '#007BFF',
-  },
-});
-
-export default styles;
